+++ startup/GafferAppleseedUI/appleseedExamples.py
@@ -1,24 +1,8 @@
 import Gaffer
 import GafferUI
-import GafferDispatch
-import GafferImage
 import GafferScene
 import GafferOSL
 import GafferAppleseed
-
-GafferUI.Examples.registerExample(
-	"Rendering/Wedge Tests",
-	"$GAFFER_ROOT/resources/examples/rendering/wedgeTests.gfr",
-	description = "Demonstrates how to use the Wedge node to render shader wedge tests.",
-	notableNodes = [
-		GafferDispatch.Wedge,
-		GafferDispatch.SystemCommand,
-		GafferImage.Text,
-		GafferScene.Outputs,
-		Gaffer.ContextVariables,
-		Gaffer.Expression
-	]
-)
 
 GafferUI.Examples.registerExample(
 	"Rendering/Anamorphic Camera",